add helper for hard-reseting files in r/w of union
Because of the nature of the beast, you can't unlink files in the underlying
r/w filesystem of a unionfs. In rare cases where starting fresh is a viable
case, system reset is required and a list of file system entries to be
removed must be processed before unionfs is mounted.

Misuse can damage your file systems and lead to continuous kernel panics.
Use with caution!